requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
SQLAlchemy==2.0.36
aiosqlite==0.20.0
httpx==0.27.2
openai==1.51.2
python-telegram-bot==21.4
stripe==10.9.0
